`Added conversion feature to PublicCode model and related controllers and views`

+++ database/migrations/2026_02_26_152359_create_measurement_units_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('measurement_units', function (Blueprint $table) {
             $table->id();
             $table->foreignId('measurement_id')->references('id')->on('public_code')->onDelete('cascade');
-            $table->float('conversion')->nullable();
+            $table->decimal('conversion', 12, 6)->nullable();
             $table->timestamps();
         });
     }
